move Player stuff to a module
start Movable interface
